Forgot to check in the last changes to netlib.c removing include of hist.h amd fixing set_hist_attributes to deal with uint64_t for histogram counters.
Stephen Burger



git-svn-id: http://www.netperf.org/svn/netperf4@106 44b83d64-6b03-0410-99c4-fe3b3a721505

+++ trunk/src/netlib.c
@@ -137,7 +137,6 @@ extern FILE * where;
 extern test_hash_t test_hash[TEST_HASH_BUCKETS];
 extern tset_hash_t test_set_hash[TEST_SET_HASH_BUCKETS];
 
-#include "hist.h"
 
 HIST
 HIST_new(void){
@@ -258,7 +257,7 @@ HIST_add(register HIST h, int time_delta)
 
 
 static xmlAttrPtr
-set_hist_attribute(xmlNodePtr hist, char *name, int *row)
+set_hist_attribute(xmlNodePtr hist, char *name, uint64_t *row)
 {
   int         i,j;
   xmlAttrPtr  ap   = NULL;
